Added admin, add product to sqlite database

+++ PosTED/app/src/main/java/com/example/posted/login/User.java
@@ -6,7 +6,7 @@ public class User {
     private String username;
     private String password;
 
-    public User(int id, String username, String password) {
+    public User(long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
